463: refactor lqr to use master/worker convention
