+++ src/start.js
@@ -30,7 +30,8 @@ import { updateNavMesh } from './interpret/graphToTriangulation';
 import { VALUE_OF_CAP } from './plan/constants';
 import { getShortestPolypointPath } from './plan/astar';
 import { drawEnemyPath, drawAllyPath } from './draw/triangulation';
-import { getLocalGoalStateFromPath, getLQRAccelerationMultipliers } from './control/lqr';
+import { getLocalGoalStateFromPath } from './control/lqr';
+import { getLQRAccelerationMultipliers } from './control/lqr.master';
 import { funnelPolypoints } from './plan/funnel';
 import { isCached } from './cache/cache';
 import { updateCache } from './cache/save';
